fix cart backend

+++ backend/services/cart/index.js
@@ -1,10 +1,11 @@
-const { deleteById } = require('./delete');
+const { deleteById, deleteByUserId } = require('./delete');
 const { readAll, readByUserId, readById } = require('./read');
-const { addItem, checkout, deleteItem, removeAllItems, updateById } = require('./update');
+const { addItem, checkout, deleteItem, removeAllItems } = require('./update');
 const { create } = require('./create');
 
 module.exports = {
     deleteById,
+    deleteByUserId,
     deleteItem,
     removeAllItems,
     readAll,
@@ -12,6 +13,5 @@ module.exports = {
     addItem,
     checkout,
     create,
-    updateById,
     readById,
 };
